requirements.txt
streamlit
google-generativeai
PyPDF2
python-docx
scikit-learn
numpy
